% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.data.R
\name{gen.data}
\alias{gen.data}
\title{Generates standard normal data X size nxp.
If cov is provided, it is used to scale each row of X.
Coefficients are generated from uniform(0,1).
If sparsity is > 0, then some entries of beta are zero-ed out.
If n.groups is provided, then n.groups number of groups are randomly generated.
The groups are returned in 0-index!}
\usage{
gen.data(n, p, cov = NA, sparsity = 0, n.groups = NA, seed = NA)
}
\description{
Generates standard normal data X size nxp.
If cov is provided, it is used to scale each row of X.
Coefficients are generated from uniform(0,1).
If sparsity is > 0, then some entries of beta are zero-ed out.
If n.groups is provided, then n.groups number of groups are randomly generated.
The groups are returned in 0-index!
}
