% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{newton_solver}
\alias{newton_solver}
\title{Solves the solution for the equation (w.r.t. \f$x\f$):
\f\link[=
     minimize \\frac{1}{2} x^\\top L x - x^\\top v 
         + l_1 ||x||_2 + \\frac{l_2}{2} ||x||_2^2
\\f]{ minimize \frac{1}{2} x^\top L x - x^\top v + l_1 ||x||_2 + \frac{l_2}{2} ||x||_2^2 \f}}
\usage{
newton_solver(L, v, l1, l2, tol, max_iters)
}
\arguments{
\item{L}{vector representing a diagonal PSD matrix.
Must have max(L + s) > 0.
L.size() <= buffer1.size().}

\item{v}{any vector.}

\item{l1}{L2-norm penalty. Must be >= 0.}

\item{l2}{L2 penalty. Must be >= 0.}

\item{tol}{Newton's method tolerance of closeness to 0.}

\item{max_iters}{maximum number of iterations of Newton's method.}

\item{x}{solution vector.}

\item{iters}{number of Newton's method iterations taken.}

\item{buffer1}{any vector with L.size() <= buffer1.size().}

\item{buffer2}{any vector with L.size() <= buffer2.size().}
}
\description{
Solves the solution for the equation (w.r.t. \f$x\f$):
\f\link[=
     minimize \\frac{1}{2} x^\\top L x - x^\\top v 
         + l_1 ||x||_2 + \\frac{l_2}{2} ||x||_2^2
\\f]{ minimize \frac{1}{2} x^\top L x - x^\top v + l_1 ||x||_2 + \frac{l_2}{2} ||x||_2^2 \f}
}
