% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{objective}
\alias{objective}
\title{Computes the group-lasso objective.}
\usage{
objective(A, r, groups, group_sizes, alpha, penalty, lmda, beta)
}
\arguments{
\item{A}{any square (p, p) matrix.}

\item{r}{any vector (p,).}

\item{groups}{vector defining group beginning indices.}

\item{group_sizes}{vector defining group sizes.}

\item{alpha}{elastic net proportion.}

\item{penalty}{penalty factor for each group.}

\item{lmda}{group-lasso regularization.}

\item{beta}{coefficient vector.}
}
\description{
Computes the group-lasso objective.
}
